# backend/internal/handlers/admin.go
package handlers

import (
	"errors"
	"log"
	"strconv"

	"github.com/afa/blueprint/backend/internal/domain"
	"github.com/afa/blueprint/backend/pkg/config"
	"github.com/gofiber/fiber/v2"
)

type AdminHandler struct {
	users       domain.UserRepository
	banners     domain.BannerRepository
	linktree    domain.LinktreeRepository
	brandKit    domain.BrandKitRepository
	emailGroups domain.EmailGroupRepository
	emailSubs   domain.EmailSubscriptionRepository
	userGroups  domain.UserGroupRepository
	storage     domain.Storage
	cfg         *config.Config
}

func NewAdminHandler(
	users domain.UserRepository,
	banners domain.BannerRepository,
	linktree domain.LinktreeRepository,
	brandKit domain.BrandKitRepository,
	emailGroups domain.EmailGroupRepository,
	emailSubs domain.EmailSubscriptionRepository,
	userGroups domain.UserGroupRepository,
	storage domain.Storage,
	cfg *config.Config,
) *AdminHandler {
	return &AdminHandler{
		users:       users,
		banners:     banners,
		linktree:    linktree,
		brandKit:    brandKit,
		emailGroups: emailGroups,
		emailSubs:   emailSubs,
		userGroups:  userGroups,
		storage:     storage,
		cfg:         cfg,
	}
}

// Upload handles generic admin file uploads. The frontend sends a multipart
// form with a `file` field and an optional `prefix` (subfolder) field. Only
// a fixed allowlist of prefixes is accepted.
// Upload godoc
// @Summary     Upload a file (admin)
// @Tags        Admin
// @Accept      multipart/form-data
// @Produce     json
// @Param       file formData file true "File to upload"
// @Param       prefix formData string false "Upload prefix (uploads/covers/banners/linktree/brand-kit/products)"
// @Success     200 {object} map[string]interface{}
// @Security    BearerAuth
// @Router      /admin/upload [post]
func (h *AdminHandler) Upload(c *fiber.Ctx) error {
	file, err := c.FormFile("file")
	if err != nil {
		return fiber.NewError(fiber.StatusBadRequest, "file is required")
	}

	prefix := c.FormValue("prefix", "uploads")
	allowed := map[string]bool{
		"uploads":   true,
		"covers":    true,
		"banners":   true,
		"linktree":  true,
		"brand-kit": true,
		"products":  true,
	}
	if !allowed[prefix] {
		return fiber.NewError(fiber.StatusBadRequest, "invalid prefix")
	}

	url, err := UploadFormFile(c.Context(), h.storage, file, prefix)
	if err != nil {
		if errors.Is(err, domain.ErrInvalidInput) {
			return fiber.NewError(fiber.StatusBadRequest, "invalid upload")
		}
		log.Printf("admin.Upload: upload failed (prefix=%s, file=%q): %v", prefix, file.Filename, err)
		return fiber.NewError(fiber.StatusInternalServerError, "upload failed")
	}
	return c.JSON(fiber.Map{"url": url})
}

// pagination helpers
func paginate(c *fiber.Ctx) (page, limit, offset int) {
	page, _ = strconv.Atoi(c.Query("page", "1"))
	limit, _ = strconv.Atoi(c.Query("limit", "20"))
	if page < 1 {
		page = 1
	}
	if limit < 1 {
		limit = 20
	}
	offset = (page - 1) * limit
	return
}

// ---- Users ----

// ListUsers godoc
// @Summary     List all users (admin)
// @Tags        Admin
// @Produce     json
// @Security    BearerAuth
// @Router      /admin/users [get]
func (h *AdminHandler) ListUsers(c *fiber.Ctx) error {
	page, limit, offset := paginate(c)
	users, total, err := h.users.List(c.Context(), offset, limit)
	if err != nil {
		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
	}
	return c.JSON(fiber.Map{"data": users, "total": total, "page": page, "limit": limit})
}

// UpdateUserRole godoc
// @Summary     Update user role (admin)
// @Tags        Admin
// @Accept      json
// @Produce     json
// @Security    BearerAuth
// @Router      /admin/users/{id}/role [put]
func (h *AdminHandler) UpdateUserRole(c *fiber.Ctx) error {
	id := c.Params("id")
	var body struct {
		Role string `json:"role"`
	}
	if err := c.BodyParser(&body); err != nil {
		return fiber.NewError(fiber.StatusBadRequest, err.Error())
	}
	switch body.Role {
	case "admin", "operator", "user":
	default:
		return fiber.NewError(fiber.StatusBadRequest, "invalid role")
	}
	user, err := h.users.FindByID(c.Context(), id)
	if err != nil {
		return fiber.NewError(fiber.StatusNotFound, "user not found")
	}
	user.Role = body.Role
	if err := h.users.Update(c.Context(), user); err != nil {
		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
	}
	return c.JSON(user)
}

// DeleteUser godoc
// @Summary     Delete a user (admin)
// @Tags        Admin
// @Security    BearerAuth
// @Router      /admin/users/{id} [delete]
func (h *AdminHandler) DeleteUser(c *fiber.Ctx) error {
	id := c.Params("id")
	if err := h.users.Delete(c.Context(), id); err != nil {
		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
	}
	return c.SendStatus(fiber.StatusNoContent)
}

// ---- Banners ----

// ListBanners godoc
// @Summary     List all banners (admin)
// @Tags        Admin
// @Produce     json
// @Security    BearerAuth
// @Router      /admin/banners [get]
func (h *AdminHandler) ListBanners(c *fiber.Ctx) error {
	banners, err := h.banners.List(c.Context(), false)
	if err != nil {
		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
	}
	return c.JSON(fiber.Map{"data": banners})
}

// CreateBanner godoc
// @Summary     Create a banner (admin)
// @Tags        Admin
// @Accept      json
// @Produce     json
// @Security    BearerAuth
// @Router      /admin/banners [post]
func (h *AdminHandler) CreateBanner(c *fiber.Ctx) error {
	var b domain.Banner
	if err := c.BodyParser(&b); err != nil {
		return fiber.NewError(fiber.StatusBadRequest, err.Error())
	}
	if err := h.banners.Create(c.Context(), &b); err != nil {
		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
	}
	return c.Status(fiber.StatusCreated).JSON(b)
}

// UpdateBanner godoc
// @Summary     Update a banner (admin)
// @Tags        Admin
// @Accept      json
// @Produce     json
// @Security    BearerAuth
// @Router      /admin/banners/{id} [put]
func (h *AdminHandler) UpdateBanner(c *fiber.Ctx) error {
	var b domain.Banner
	if err := c.BodyParser(&b); err != nil {
		return fiber.NewError(fiber.StatusBadRequest, err.Error())
	}
	b.ID = c.Params("id")
	if err := h.banners.Update(c.Context(), &b); err != nil {
		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
	}
	return c.JSON(b)
}

// DeleteBanner godoc
// @Summary     Delete a banner (admin)
// @Tags        Admin
// @Security    BearerAuth
// @Router      /admin/banners/{id} [delete]
func (h *AdminHandler) DeleteBanner(c *fiber.Ctx) error {
	if err := h.banners.Delete(c.Context(), c.Params("id")); err != nil {
		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
	}
	return c.SendStatus(fiber.StatusNoContent)
}

// ---- Linktree ----

// ListLinktree godoc
// @Summary     List linktree items (admin)
// @Tags        Admin
// @Produce     json
// @Security    BearerAuth
// @Router      /admin/linktree [get]
func (h *AdminHandler) ListLinktree(c *fiber.Ctx) error {
	items, err := h.linktree.List(c.Context(), false)
	if err != nil {
		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
	}
	return c.JSON(fiber.Map{"data": items})
}

// CreateLinktreeItem godoc
// @Summary     Create a linktree item (admin)
// @Tags        Admin
// @Accept      json
// @Produce     json
// @Security    BearerAuth
// @Router      /admin/linktree [post]
func (h *AdminHandler) CreateLinktreeItem(c *fiber.Ctx) error {
	var item domain.LinktreeItem
	if err := c.BodyParser(&item); err != nil {
		return fiber.NewError(fiber.StatusBadRequest, err.Error())
	}
	if err := h.linktree.Create(c.Context(), &item); err != nil {
		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
	}
	return c.Status(fiber.StatusCreated).JSON(item)
}

// UpdateLinktreeItem godoc
// @Summary     Update a linktree item (admin)
// @Tags        Admin
// @Accept      json
// @Produce     json
// @Security    BearerAuth
// @Router      /admin/linktree/{id} [put]
func (h *AdminHandler) UpdateLinktreeItem(c *fiber.Ctx) error {
	var item domain.LinktreeItem
	if err := c.BodyParser(&item); err != nil {
		return fiber.NewError(fiber.StatusBadRequest, err.Error())
	}
	item.ID = c.Params("id")
	if err := h.linktree.Update(c.Context(), &item); err != nil {
		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
	}
	return c.JSON(item)
}

// DeleteLinktreeItem godoc
// @Summary     Delete a linktree item (admin)
// @Tags        Admin
// @Security    BearerAuth
// @Router      /admin/linktree/{id} [delete]
func (h *AdminHandler) DeleteLinktreeItem(c *fiber.Ctx) error {
	if err := h.linktree.Delete(c.Context(), c.Params("id")); err != nil {
		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
	}
	return c.SendStatus(fiber.StatusNoContent)
}

// ReorderLinktree godoc
// @Summary     Reorder linktree items (admin)
// @Tags        Admin
// @Accept      json
// @Security    BearerAuth
// @Router      /admin/linktree/reorder [put]
func (h *AdminHandler) ReorderLinktree(c *fiber.Ctx) error {
	var items []domain.LinktreeItem
	if err := c.BodyParser(&items); err != nil {
		return fiber.NewError(fiber.StatusBadRequest, err.Error())
	}
	if err := h.linktree.Reorder(c.Context(), items); err != nil {
		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
	}
	return c.SendStatus(fiber.StatusNoContent)
}

// ---- Brand Kit ----

// GetBrandKit godoc
// @Summary     Get brand kit (admin)
// @Tags        Admin
// @Produce     json
// @Security    BearerAuth
// @Router      /admin/brand-kit [get]
func (h *AdminHandler) GetBrandKit(c *fiber.Ctx) error {
	bk, err := h.brandKit.Get(c.Context())
	if err != nil {
		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
	}
	return c.JSON(bk)
}

// UpsertBrandKit godoc
// @Summary     Update brand kit (admin)
// @Tags        Admin
// @Accept      json
// @Produce     json
// @Security    BearerAuth
// @Router      /admin/brand-kit [put]
func (h *AdminHandler) UpsertBrandKit(c *fiber.Ctx) error {
	var bk domain.BrandKit
	if err := c.BodyParser(&bk); err != nil {
		return fiber.NewError(fiber.StatusBadRequest, err.Error())
	}
	if err := h.brandKit.Upsert(c.Context(), &bk); err != nil {
		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
	}
	return c.JSON(bk)
}

// ---- Email Groups ----

// ListEmailGroups godoc
// @Summary     List email groups (admin)
// @Tags        Admin
// @Produce     json
// @Security    BearerAuth
// @Router      /admin/email-groups [get]
func (h *AdminHandler) ListEmailGroups(c *fiber.Ctx) error {
	groups, err := h.emailGroups.List(c.Context())
	if err != nil {
		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
	}
	return c.JSON(fiber.Map{"data": groups})
}

// CreateEmailGroup godoc
// @Summary     Create an email group (admin)
// @Tags        Admin
// @Accept      json
// @Produce     json
// @Security    BearerAuth
// @Router      /admin/email-groups [post]
func (h *AdminHandler) CreateEmailGroup(c *fiber.Ctx) error {
	var g domain.EmailGroup
	if err := c.BodyParser(&g); err != nil {
		return fiber.NewError(fiber.StatusBadRequest, err.Error())
	}
	if g.Name == "" {
		return c.Status(400).JSON(fiber.Map{"error": "name is required"})
	}
	if err := h.emailGroups.Create(c.Context(), &g); err != nil {
		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
	}
	return c.Status(fiber.StatusCreated).JSON(g)
}

// DeleteEmailGroup godoc
// @Summary     Delete an email group (admin)
// @Tags        Admin
// @Security    BearerAuth
// @Router      /admin/email-groups/{id} [delete]
func (h *AdminHandler) DeleteEmailGroup(c *fiber.Ctx) error {
	if err := h.emailGroups.Delete(c.Context(), c.Params("id")); err != nil {
		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
	}
	return c.JSON(fiber.Map{"deleted": true})
}

// ---- Email Subscriptions ----

// ListSubscribers godoc
// @Summary     List subscribers in a group (admin)
// @Tags        Admin
// @Produce     json
// @Security    BearerAuth
// @Router      /admin/email-groups/{id}/subscribers [get]
func (h *AdminHandler) ListSubscribers(c *fiber.Ctx) error {
	subs, err := h.emailSubs.ListByGroup(c.Context(), c.Params("id"))
	if err != nil {
		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
	}
	return c.JSON(fiber.Map{"data": subs})
}

// AddEmailSubscription godoc
// @Summary     Add an email subscription (admin)
// @Tags        Admin
// @Accept      json
// @Produce     json
// @Security    BearerAuth
// @Router      /admin/email-subscriptions [post]
func (h *AdminHandler) AddEmailSubscription(c *fiber.Ctx) error {
	var sub domain.EmailSubscription
	if err := c.BodyParser(&sub); err != nil {
		return fiber.NewError(fiber.StatusBadRequest, err.Error())
	}
	if err := h.emailSubs.Add(c.Context(), &sub); err != nil {
		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
	}
	return c.Status(fiber.StatusCreated).JSON(sub)
}

// DeactivateEmailSubscription godoc
// @Summary     Deactivate an email subscription (admin)
// @Tags        Admin
// @Security    BearerAuth
// @Router      /admin/email-subscriptions/{email}/deactivate [put]
func (h *AdminHandler) DeactivateEmailSubscription(c *fiber.Ctx) error {
	email := c.Params("email")
	if err := h.emailSubs.Deactivate(c.Context(), email); err != nil {
		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
	}
	return c.SendStatus(fiber.StatusNoContent)
}

// ---- User Groups ----

// ListUserGroups godoc
// @Summary     List user groups (admin)
// @Tags        Admin
// @Produce     json
// @Security    BearerAuth
// @Router      /admin/user-groups [get]
func (h *AdminHandler) ListUserGroups(c *fiber.Ctx) error {
	groups, err := h.userGroups.List(c.Context())
	if err != nil {
		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
	}
	return c.JSON(fiber.Map{"data": groups})
}

// CreateUserGroup godoc
// @Summary     Create a user group (admin)
// @Tags        Admin
// @Accept      json
// @Produce     json
// @Security    BearerAuth
// @Router      /admin/user-groups [post]
func (h *AdminHandler) CreateUserGroup(c *fiber.Ctx) error {
	var g domain.UserGroup
	if err := c.BodyParser(&g); err != nil {
		return fiber.NewError(fiber.StatusBadRequest, err.Error())
	}
	if err := h.userGroups.Create(c.Context(), &g); err != nil {
		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
	}
	return c.Status(fiber.StatusCreated).JSON(g)
}

// DeleteUserGroup godoc
// @Summary     Delete a user group (admin)
// @Tags        Admin
// @Security    BearerAuth
// @Router      /admin/user-groups/{id} [delete]
func (h *AdminHandler) DeleteUserGroup(c *fiber.Ctx) error {
	if err := h.userGroups.Delete(c.Context(), c.Params("id")); err != nil {
		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
	}
	return c.SendStatus(fiber.StatusNoContent)
}